requirements.txt
openai-whisper==20250625
sounddevice==0.5.3
scipy==1.16.3
pynput==1.8.1
pyperclip==1.11.0
torch==2.9.1
torchaudio==2.9.1
numpy==2.3.5
